pyproject.toml
[project]
name = "breviori"
version = "1.0.0"
readme = "README.md"
requires-python = "^3.12.0"
urls = { homepage = "https://apis.vicit.studio" }
description = "Web App simples para encurta urls."
authors = [
    {name = "afiovinicius",email = "afiovinicius@gmail.com"}
]

dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "redis[hiredis] (>=5.2.1,<6.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "distutils-extra-python (>=2.39.2,<3.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "setuptools (>=79.0.1,<80.0.0)",
    "strawberry-graphql[fastapi] (>=0.266.0,<0.267.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
blue = "^0.9.1"
ruff = "^0.11.7"
httpx = "^0.28.1"
isort = "^6.0.1"
ignr = "^2.2"
python-dotenv = "^1.1.0"

[tool.ruff]
line-length = 80
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 80
extend_skip = ['migrations']

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff && isort --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn app.main:app --reload'

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
